+++ src/subscriber.rs
@@ -42,7 +42,7 @@ pub fn configure_subscriber<T: EventFormatter + Send + Sync + 'static>(
             .with(PrimaJsonStorage::default())
             .with(PrimaFormattingLayer::new(
                 _config.service.clone(),
-                _config.country.map(|country| country.to_string()),
+                _config.country.to_string(),
                 _config.env.to_string(),
                 &std::io::stdout,
                 _config.json_formatter,
@@ -86,7 +86,7 @@ impl Drop for Uninstall {
 /// Information about the current app context like name or environment
 pub struct ContextInfo<'a> {
     pub(crate) app_name: &'a str,
-    pub(crate) country: &'a Option<&'a str>,
+    pub(crate) country: &'a str,
     pub(crate) environment: &'a str,
 }
 
@@ -95,7 +95,7 @@ impl<'a> ContextInfo<'a> {
         self.app_name
     }
 
-    pub fn country(&self) -> &'a Option<&'a str> {
+    pub fn country(&self) -> &'a str {
         self.country
     }
 
@@ -104,7 +104,7 @@ impl<'a> ContextInfo<'a> {
     }
 }
 
-pub struct NopEventFormatter {}
+pub struct NopEventFormatter;
 
 impl EventFormatter for NopEventFormatter {
     fn format_event<S>(
